Simply test against CLOCK_REALTIME and CLOCK_MONOTONIC.
Just like GNU Hurd and OpenBSD, Nuxi CloudABI[1] implements
clock_gettime(), but cannot define _POSIX_TIMERS. The reason for this is
that CloudABI cannot implement the timer_*() interface, as it does not
provide full UNIX signal handling support.

Instead of making the checks in this source file even more complex,
simply fall back to testing against CLOCK_REALTIME. I'd say it would be
very unlikely for an implementation to provide CLOCK_REALTIME, but not
offer a working clock_gettime() function that can operate on it.

[1] Nuxi CloudABI: https://github.com/NuxiNL/cloudlibc