Added casts of date/time components to uint32_t to silence MSVC warnings.
The date and time values handled by Boost.Log are not supposed to exceed
uint32_t range most of the time. The possible exceptions are date/time
durations, which are currently handled by the same code path as the
date/time points. If durations become an issue, the code should be
reworked so that durations are handled separately by the formater.
Converting all of the implementation to uint64_t won't help since the
decomposed date/time is further converted to std::tm, which is unlikely to
have uint64_t-wide components.