Use SFINAE to limit attribute set iterator comparison, copy and assignment.
This makes the code more SFINAE-friendly. It also changes the converting
constructor and assignment operator to only work as converting constructor and
assignment and let the compiler generate the implicit copy constructor and
assignment operator. This should silence bogus clang-10 warnings about
deprecated generation of an implicit copy assignment operator in presence of
a copy constructor. (The warning is bogus because there already is an explicitly
defined assignment operator, which works as a copy assignment as well.)

Closes https://github.com/boostorg/log/issues/118.