Added checking for IP version in addresses in syslog backend.
The syslog sink backend now verifies the IP version of the local and target
addresses set by user. The addresses must have the same IP version as was
specified in the ip_version named parameter on the sink backend construction.
When an address is obtained as a result of host name resolution, only addresses
with matching IP version are considered.

This should protect against using local and target addresses with mismatching
IP versions, which results in errors on sending datagrams.

Fixes https://github.com/boostorg/log/issues/119.