Fixed file counter in the target filename generated by text file backend.
The file counter gets incremented when a new file is opened. Since the
target filename was generated later, on file rotation, the target filename
got the next value of the counter. This caused some values of the counter
being skipped when the application restarts and scans for the previous
log files.

Fixes https://github.com/boostorg/log/issues/125.