Resolved a possible conflict between flush and run in asynchronous_sink.
If run is called in a user's thread, and flush is called before the user's
thread enters run, flush would mark the sink frontend as being used by
another record feeding thread and run would exit with an exception.

To mitigate this, we now only throw exception when either run or
feed_records is being called in multiple threads, but not flush. When flush
is in progress, run and feed_records will block until the flush is
complete. This ensures that the sink backend is only used from one thread
at a time.

Additionally, stop now explicitly doesn't wait for the ongoing record
feeding operation to complete. It is generally not possible to ensure that
the record feeding thread has left (or won't enter) the feeding operation
after stop returns. When the internal feeding thread is used we ensure
that the thread is done by joining it. When user's feeding thread is used
it is the user's responsibility to ensure that the thread won't use
the frontend past stop. Joining the thread is one possible way to do it.

Fixes https://github.com/boostorg/log/issues/131.