Corrected a formal race in the threadsafe_queue implementation.
The pointer to the next node can be read and written concurrently
by push and try_pop methods. This does not impose a real problem
as long as pointer-sized loads and stores generated by the compiler
are atomic on a given target platform, which it typically is.
No memory ordering constraints are needed in this context.

To guarantee atomicity, use Boost.Atomic for these pointers.

Fixes https://github.com/boostorg/log/issues/139.