Resolved Boost.Build warnings and fixed building docs on Windows.
Boost.Build was complaining about reserved characters used in boost.url.prefix,
which is presumably a colon. Wrapped feature values in double quotes to silence
that.

In the gen-references action, there was a commented line, which caused errors
on Windows. Presumably, because the cmd shell did not interpret # as a comment.
Removed the comment to resolve the issue.