Restored C++ versions to build on MacOS and disabled catching signals.
Removing newer C++ versions from MacOS CI job didn't help with job hanging,
so we're restoring the complete C++ version list.

As another attempt to work around the problem, disable signal handling
and debugger attaching in Boost.Test. This is in case if there is a genuine
problem with a test crashing and Boost.Test preventing the process from
terminating.