Added a dummy export from boost_log_setup if settings parsers are disabled.
When settings parsers were disabled by a config macro, boost_log_setup would
be empty, and MSVC linker would produce no library at all. This broke
installing Boost.Log, as Boost.Build would expect a library file to be present.

We're adding a dummy export from the library as a workaround, so that the
library is always created, even if empty. This workaround should be considered
temporary, until we can disable the library proper after a Boost.Build bug
is fixed:

https://github.com/bfgroup/b2/issues/104

Closes https://github.com/boostorg/log/issues/164.
Closes https://github.com/boostorg/log/pull/165.