Improved performance of SSSE3/AVX2 implementations of dump manipulator.
Use byte shuffle instead of arithmetics and masking to convert half-bytes
to characters.

Since the new algorithm more heavilly relies on pshufb, and it is slow
on some older Intel Atom CPUs, retain the old implementation and use it
when we know pshufb is slow. We might drop it eventually.