Removed auto-linking Boost.System and Boost.DateTime.
Boost.System and Boost.DateTime are header-only now. We used to include
Boost.System config header to auto-link with it (which it no longer does)
and manually auto-link Boost.DateTime. The latter caused a problem with
partial Boost checkout and build/install, when the enabled dependencies were
derived from the Boost.Log Jamfile, which does not contain a dependency on
DateTime.

Closes https://github.com/boostorg/log/issues/166.