Don't rotate the empty log file if the message size exceeds rotation_size.
In text_file_backend, we try to avoid exceeding the rotation_size limit when we
open a new file, potentially in append mode. To do so, we check the combined
size of the opened file and the formatted log record to write against this
limit. It was possible that the log record alone exceeds the limit, which makes
this check always fail and cause an infinite loop of rotating files.

To avoid this, don't perform the check if the file size is 0, which is the case
when a new file is created. In this case, we will write the log record (thus
exceeding the rotation_size limit) and then rotate the file upon the next
log record we write.

Closes https://github.com/boostorg/log/issues/209.